+++ src/spl/token/instructions.py
@@ -28,7 +28,7 @@ class AuthorityType(IntEnum):
 
 
 # Instruction Params
-@deprecated("InitializeMintParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeMintParams is deprecated; use spl.token.models instead.")
 class InitializeMintParams(NamedTuple):
     """Initialize token mint transaction params."""
 
@@ -44,7 +44,7 @@ class InitializeMintParams(NamedTuple):
     """The freeze authority/multisignature of the mint."""
 
 
-@deprecated("InitializeMint2Params is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeMint2Params is deprecated; use spl.token.models instead.")
 class InitializeMint2Params(NamedTuple):
     """Initialize token mint transaction params without Rent sysvar."""
 
@@ -60,7 +60,7 @@ class InitializeMint2Params(NamedTuple):
     """The freeze authority/multisignature of the mint."""
 
 
-@deprecated("InitializeAccountParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeAccountParams is deprecated; use spl.token.models instead.")
 class InitializeAccountParams(NamedTuple):
     """Initialize token account transaction params."""
 
@@ -74,7 +74,7 @@ class InitializeAccountParams(NamedTuple):
     """Owner of the new account."""
 
 
-@deprecated("InitializeAccount2Params is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeAccount2Params is deprecated; use spl.token.models instead.")
 class InitializeAccount2Params(NamedTuple):
     """Initialize token account transaction params with owner in instruction data."""
 
@@ -88,7 +88,7 @@ class InitializeAccount2Params(NamedTuple):
     """Owner of the new account."""
 
 
-@deprecated("InitializeAccount3Params is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeAccount3Params is deprecated; use spl.token.models instead.")
 class InitializeAccount3Params(NamedTuple):
     """Initialize token account transaction params with owner in instruction data and no Rent sysvar."""
 
@@ -102,7 +102,7 @@ class InitializeAccount3Params(NamedTuple):
     """Owner of the new account."""
 
 
-@deprecated("InitializeMultisigParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeMultisigParams is deprecated; use spl.token.models instead.")
 class InitializeMultisigParams(NamedTuple):
     """Initialize multisig token account transaction params."""
 
@@ -116,7 +116,7 @@ class InitializeMultisigParams(NamedTuple):
     """Addresses of multisig signers."""
 
 
-@deprecated("InitializeMultisig2Params is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeMultisig2Params is deprecated; use spl.token.models instead.")
 class InitializeMultisig2Params(NamedTuple):
     """Initialize multisig token account transaction params without Rent sysvar."""
 
@@ -130,7 +130,7 @@ class InitializeMultisig2Params(NamedTuple):
     """Addresses of multisig signers."""
 
 
-@deprecated("TransferParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("TransferParams is deprecated; use spl.token.models instead.")
 class TransferParams(NamedTuple):
     """Transfer token transaction params."""
 
@@ -148,7 +148,7 @@ class TransferParams(NamedTuple):
     """Signing accounts if `owner` is a multiSig."""
 
 
-@deprecated("ApproveParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("ApproveParams is deprecated; use spl.token.models instead.")
 class ApproveParams(NamedTuple):
     """Approve token transaction params."""
 
@@ -166,7 +166,7 @@ class ApproveParams(NamedTuple):
     """Signing accounts if `owner` is a multiSig."""
 
 
-@deprecated("RevokeParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("RevokeParams is deprecated; use spl.token.models instead.")
 class RevokeParams(NamedTuple):
     """Revoke token transaction params."""
 
@@ -180,7 +180,7 @@ class RevokeParams(NamedTuple):
     """Signing accounts if `owner` is a multiSig."""
 
 
-@deprecated("SetAuthorityParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("SetAuthorityParams is deprecated; use spl.token.models instead.")
 class SetAuthorityParams(NamedTuple):
     """Set token authority transaction params."""
 
@@ -198,7 +198,7 @@ class SetAuthorityParams(NamedTuple):
     """New authority of the account."""
 
 
-@deprecated("MintToParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("MintToParams is deprecated; use spl.token.models instead.")
 class MintToParams(NamedTuple):
     """Mint token transaction params."""
 
@@ -216,7 +216,7 @@ class MintToParams(NamedTuple):
     """Signing accounts if `mint_authority` is a multiSig."""
 
 
-@deprecated("BurnParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("BurnParams is deprecated; use spl.token.models instead.")
 class BurnParams(NamedTuple):
     """Burn token transaction params."""
 
@@ -234,7 +234,7 @@ class BurnParams(NamedTuple):
     """Signing accounts if `owner` is a multiSig"""
 
 
-@deprecated("CloseAccountParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("CloseAccountParams is deprecated; use spl.token.models instead.")
 class CloseAccountParams(NamedTuple):
     """Close token account transaction params."""
 
@@ -250,7 +250,7 @@ class CloseAccountParams(NamedTuple):
     """Signing accounts if `owner` is a multiSig"""
 
 
-@deprecated("FreezeAccountParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("FreezeAccountParams is deprecated; use spl.token.models instead.")
 class FreezeAccountParams(NamedTuple):
     """Freeze token account transaction params."""
 
@@ -266,7 +266,7 @@ class FreezeAccountParams(NamedTuple):
     """Signing accounts if `authority` is a multiSig"""
 
 
-@deprecated("ThawAccountParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("ThawAccountParams is deprecated; use spl.token.models instead.")
 class ThawAccountParams(NamedTuple):
     """Thaw token account transaction params."""
 
@@ -282,7 +282,7 @@ class ThawAccountParams(NamedTuple):
     """Signing accounts if `authority` is a multiSig"""
 
 
-@deprecated("TransferCheckedParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("TransferCheckedParams is deprecated; use spl.token.models instead.")
 class TransferCheckedParams(NamedTuple):
     """TransferChecked token transaction params."""
 
@@ -304,7 +304,7 @@ class TransferCheckedParams(NamedTuple):
     """Signing accounts if `owner` is a multiSig."""
 
 
-@deprecated("ApproveCheckedParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("ApproveCheckedParams is deprecated; use spl.token.models instead.")
 class ApproveCheckedParams(NamedTuple):
     """ApproveChecked token transaction params."""
 
@@ -326,7 +326,7 @@ class ApproveCheckedParams(NamedTuple):
     """Signing accounts if `owner` is a multiSig."""
 
 
-@deprecated("MintToCheckedParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("MintToCheckedParams is deprecated; use spl.token.models instead.")
 class MintToCheckedParams(NamedTuple):
     """MintToChecked token transaction params."""
 
@@ -346,7 +346,7 @@ class MintToCheckedParams(NamedTuple):
     """Signing accounts if `mint_authority` is a multiSig."""
 
 
-@deprecated("BurnCheckedParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("BurnCheckedParams is deprecated; use spl.token.models instead.")
 class BurnCheckedParams(NamedTuple):
     """BurnChecked token transaction params."""
 
@@ -366,7 +366,7 @@ class BurnCheckedParams(NamedTuple):
     """Signing accounts if `owner` is a multiSig"""
 
 
-@deprecated("SyncNativeParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("SyncNativeParams is deprecated; use spl.token.models instead.")
 class SyncNativeParams(NamedTuple):
     """BurnChecked token transaction params."""
 
@@ -376,7 +376,7 @@ class SyncNativeParams(NamedTuple):
     """Account to sync."""
 
 
-@deprecated("GetAccountDataSizeParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("GetAccountDataSizeParams is deprecated; use spl.token.models instead.")
 class GetAccountDataSizeParams(NamedTuple):
     """GetAccountDataSize token transaction params."""
 
@@ -386,7 +386,7 @@ class GetAccountDataSizeParams(NamedTuple):
     """Mint to calculate account size for."""
 
 
-@deprecated("InitializeImmutableOwnerParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeImmutableOwnerParams is deprecated; use spl.token.models instead.")
 class InitializeImmutableOwnerParams(NamedTuple):
     """InitializeImmutableOwner token transaction params."""
 
@@ -396,7 +396,7 @@ class InitializeImmutableOwnerParams(NamedTuple):
     """Token account to initialize immutable owner for."""
 
 
-@deprecated("InitializeTransferFeeConfigParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("InitializeTransferFeeConfigParams is deprecated; use spl.token.models instead.")
 class InitializeTransferFeeConfigParams(NamedTuple):
     """InitializeTransferFeeConfig token transaction params."""
 
@@ -414,7 +414,7 @@ class InitializeTransferFeeConfigParams(NamedTuple):
     """Maximum fee assessed on transfers."""
 
 
-@deprecated("WithdrawWithheldTokensFromAccountsParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("WithdrawWithheldTokensFromAccountsParams is deprecated; use spl.token.models instead.")
 class WithdrawWithheldTokensFromAccountsParams(NamedTuple):
     """WithdrawWithheldTokensFromAccounts token transaction params."""
 
@@ -432,7 +432,7 @@ class WithdrawWithheldTokensFromAccountsParams(NamedTuple):
     """Token accounts to withdraw withheld tokens from."""
 
 
-@deprecated("WithdrawWithheldTokensFromMintParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("WithdrawWithheldTokensFromMintParams is deprecated; use spl.token.models instead.")
 class WithdrawWithheldTokensFromMintParams(NamedTuple):
     """WithdrawWithheldTokensFromMint token transaction params."""
 
@@ -448,7 +448,7 @@ class WithdrawWithheldTokensFromMintParams(NamedTuple):
     """Signing accounts if `authority` is a multiSig."""
 
 
-@deprecated("HarvestWithheldTokensToMintParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("HarvestWithheldTokensToMintParams is deprecated; use spl.token.models instead.")
 class HarvestWithheldTokensToMintParams(NamedTuple):
     """HarvestWithheldTokensToMint token transaction params."""
 
@@ -460,7 +460,7 @@ class HarvestWithheldTokensToMintParams(NamedTuple):
     """Token accounts to harvest withheld tokens from."""
 
 
-@deprecated("AmountToUiAmountParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("AmountToUiAmountParams is deprecated; use spl.token.models instead.")
 class AmountToUiAmountParams(NamedTuple):
     """AmountToUiAmount token transaction params."""
 
@@ -472,7 +472,7 @@ class AmountToUiAmountParams(NamedTuple):
     """Amount of tokens to reformat."""
 
 
-@deprecated("UiAmountToAmountParams is deprecated and will be replaced by a Pydantic model.")
+@deprecated("UiAmountToAmountParams is deprecated; use spl.token.models instead.")
 class UiAmountToAmountParams(NamedTuple):
     """UiAmountToAmount token transaction params."""